backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
openai==1.76.0
asyncpg==0.30.0
pypdf==5.4.0
orjson==3.10.16
httpx==0.28.1
